Calculate delta values at t time with ODE solutions from ana_slvr

Calculate the delta values at t time using the ODE analytical solutions of the isotopic box model.

ANA_delta_t_Calculator(
  t,
  ODE_Constants,
  ODE_Eigenvalues,
  ODE_Eigenvectors,
  BOXES_IDs,
  ratio_standard
)

Arguments

t
Time at which the delta values are to be calculated (numeric)

ODE_Constants
Constants as determined by the analytical solver for the system of
ordinary differential equations (single column dataframe).

ODE_Eigenvalues
Eigenvalues as determined by the analytical solver for the system of
ordinary differential equations (single column dataframe).

ODE_Eigenvectors
Eigenvectors as determined by the analytical solver for the system of
ordinary differential equations (multiple columns dataframe).

BOXES_IDs
Vector of character strings with the names of the boxes
in the same order as used in ana_slvr.

ratio_standard
Isotope ratio of the reference material used to calculate the delta values.

Value

Dataframe of the delta values in all boxes at t time.

See also