Verticalizes a dataframe

Takes a dataframe with a set of different columns containing numerical values to be verticalized, returns a vertical dataframe with all variables in a single column called "VAR" together with a "VAR_TYPE" column defining the type of variable for the given row, named after the column name found in horizontal dataframe.

DF_verticalizer(df_hor, vert_col)

Arguments

df_hor
Horizontal dataframe

vert_col
Vector of the names of the columns of numerical variables to be verticalized.

Value

A vertical dataframe containing the variables to be verticalized (column "VAR", numeric) and the name of the variable (column "VAR_TYPE", character strings).

Examples

DF_verticalizer(df_hor = iris, vert_col = c("Sepal.Length", "Sepal.Width", "Petal.Length", "Petal.Width"))

#>        Species VAR     VAR_TYPE
#> 1       setosa 5.1 Sepal.Length
#> 2       setosa 4.9 Sepal.Length
#> 3       setosa 4.7 Sepal.Length
#> 4       setosa 4.6 Sepal.Length
#> 5       setosa 5.0 Sepal.Length
#> 6       setosa 5.4 Sepal.Length
#> 7       setosa 4.6 Sepal.Length
#> 8       setosa 5.0 Sepal.Length
#> 9       setosa 4.4 Sepal.Length
#> 10      setosa 4.9 Sepal.Length
#> 11      setosa 5.4 Sepal.Length
#> 12      setosa 4.8 Sepal.Length
#> 13      setosa 4.8 Sepal.Length
#> 14      setosa 4.3 Sepal.Length
#> 15      setosa 5.8 Sepal.Length
#> 16      setosa 5.7 Sepal.Length
#> 17      setosa 5.4 Sepal.Length
#> 18      setosa 5.1 Sepal.Length
#> 19      setosa 5.7 Sepal.Length
#> 20      setosa 5.1 Sepal.Length
#> 21      setosa 5.4 Sepal.Length
#> 22      setosa 5.1 Sepal.Length
#> 23      setosa 4.6 Sepal.Length
#> 24      setosa 5.1 Sepal.Length
#> 25      setosa 4.8 Sepal.Length
#> 26      setosa 5.0 Sepal.Length
#> 27      setosa 5.0 Sepal.Length
#> 28      setosa 5.2 Sepal.Length
#> 29      setosa 5.2 Sepal.Length
#> 30      setosa 4.7 Sepal.Length
#> 31      setosa 4.8 Sepal.Length
#> 32      setosa 5.4 Sepal.Length
#> 33      setosa 5.2 Sepal.Length
#> 34      setosa 5.5 Sepal.Length
#> 35      setosa 4.9 Sepal.Length
#> 36      setosa 5.0 Sepal.Length
#> 37      setosa 5.5 Sepal.Length
#> 38      setosa 4.9 Sepal.Length
#> 39      setosa 4.4 Sepal.Length
#> 40      setosa 5.1 Sepal.Length
#> 41      setosa 5.0 Sepal.Length
#> 42      setosa 4.5 Sepal.Length
#> 43      setosa 4.4 Sepal.Length
#> 44      setosa 5.0 Sepal.Length
#> 45      setosa 5.1 Sepal.Length
#> 46      setosa 4.8 Sepal.Length
#> 47      setosa 5.1 Sepal.Length
#> 48      setosa 4.6 Sepal.Length
#> 49      setosa 5.3 Sepal.Length
#> 50      setosa 5.0 Sepal.Length
#> 51  versicolor 7.0 Sepal.Length
#> 52  versicolor 6.4 Sepal.Length
#> 53  versicolor 6.9 Sepal.Length
#> 54  versicolor 5.5 Sepal.Length
#> 55  versicolor 6.5 Sepal.Length
#> 56  versicolor 5.7 Sepal.Length
#> 57  versicolor 6.3 Sepal.Length
#> 58  versicolor 4.9 Sepal.Length
#> 59  versicolor 6.6 Sepal.Length
#> 60  versicolor 5.2 Sepal.Length
#> 61  versicolor 5.0 Sepal.Length
#> 62  versicolor 5.9 Sepal.Length
#> 63  versicolor 6.0 Sepal.Length
#> 64  versicolor 6.1 Sepal.Length
#> 65  versicolor 5.6 Sepal.Length
#> 66  versicolor 6.7 Sepal.Length
#> 67  versicolor 5.6 Sepal.Length
#> 68  versicolor 5.8 Sepal.Length
#> 69  versicolor 6.2 Sepal.Length
#> 70  versicolor 5.6 Sepal.Length
#> 71  versicolor 5.9 Sepal.Length
#> 72  versicolor 6.1 Sepal.Length
#> 73  versicolor 6.3 Sepal.Length
#> 74  versicolor 6.1 Sepal.Length
#> 75  versicolor 6.4 Sepal.Length
#> 76  versicolor 6.6 Sepal.Length
#> 77  versicolor 6.8 Sepal.Length
#> 78  versicolor 6.7 Sepal.Length
#> 79  versicolor 6.0 Sepal.Length
#> 80  versicolor 5.7 Sepal.Length
#> 81  versicolor 5.5 Sepal.Length
#> 82  versicolor 5.5 Sepal.Length
#> 83  versicolor 5.8 Sepal.Length
#> 84  versicolor 6.0 Sepal.Length
#> 85  versicolor 5.4 Sepal.Length
#> 86  versicolor 6.0 Sepal.Length
#> 87  versicolor 6.7 Sepal.Length
#> 88  versicolor 6.3 Sepal.Length
#> 89  versicolor 5.6 Sepal.Length
#> 90  versicolor 5.5 Sepal.Length
#> 91  versicolor 5.5 Sepal.Length
#> 92  versicolor 6.1 Sepal.Length
#> 93  versicolor 5.8 Sepal.Length
#> 94  versicolor 5.0 Sepal.Length
#> 95  versicolor 5.6 Sepal.Length
#> 96  versicolor 5.7 Sepal.Length
#> 97  versicolor 5.7 Sepal.Length
#> 98  versicolor 6.2 Sepal.Length
#> 99  versicolor 5.1 Sepal.Length
#> 100 versicolor 5.7 Sepal.Length
#> 101  virginica 6.3 Sepal.Length
#> 102  virginica 5.8 Sepal.Length
#> 103  virginica 7.1 Sepal.Length
#> 104  virginica 6.3 Sepal.Length
#> 105  virginica 6.5 Sepal.Length
#> 106  virginica 7.6 Sepal.Length
#> 107  virginica 4.9 Sepal.Length
#> 108  virginica 7.3 Sepal.Length
#> 109  virginica 6.7 Sepal.Length
#> 110  virginica 7.2 Sepal.Length
#> 111  virginica 6.5 Sepal.Length
#> 112  virginica 6.4 Sepal.Length
#> 113  virginica 6.8 Sepal.Length
#> 114  virginica 5.7 Sepal.Length
#> 115  virginica 5.8 Sepal.Length
#> 116  virginica 6.4 Sepal.Length
#> 117  virginica 6.5 Sepal.Length
#> 118  virginica 7.7 Sepal.Length
#> 119  virginica 7.7 Sepal.Length
#> 120  virginica 6.0 Sepal.Length
#> 121  virginica 6.9 Sepal.Length
#> 122  virginica 5.6 Sepal.Length
#> 123  virginica 7.7 Sepal.Length
#> 124  virginica 6.3 Sepal.Length
#> 125  virginica 6.7 Sepal.Length
#> 126  virginica 7.2 Sepal.Length
#> 127  virginica 6.2 Sepal.Length
#> 128  virginica 6.1 Sepal.Length
#> 129  virginica 6.4 Sepal.Length
#> 130  virginica 7.2 Sepal.Length
#> 131  virginica 7.4 Sepal.Length
#> 132  virginica 7.9 Sepal.Length
#> 133  virginica 6.4 Sepal.Length
#> 134  virginica 6.3 Sepal.Length
#> 135  virginica 6.1 Sepal.Length
#> 136  virginica 7.7 Sepal.Length
#> 137  virginica 6.3 Sepal.Length
#> 138  virginica 6.4 Sepal.Length
#> 139  virginica 6.0 Sepal.Length
#> 140  virginica 6.9 Sepal.Length
#> 141  virginica 6.7 Sepal.Length
#> 142  virginica 6.9 Sepal.Length
#> 143  virginica 5.8 Sepal.Length
#> 144  virginica 6.8 Sepal.Length
#> 145  virginica 6.7 Sepal.Length
#> 146  virginica 6.7 Sepal.Length
#> 147  virginica 6.3 Sepal.Length
#> 148  virginica 6.5 Sepal.Length
#> 149  virginica 6.2 Sepal.Length
#> 150  virginica 5.9 Sepal.Length
#> 151     setosa 3.5  Sepal.Width
#> 152     setosa 3.0  Sepal.Width
#> 153     setosa 3.2  Sepal.Width
#> 154     setosa 3.1  Sepal.Width
#> 155     setosa 3.6  Sepal.Width
#> 156     setosa 3.9  Sepal.Width
#> 157     setosa 3.4  Sepal.Width
#> 158     setosa 3.4  Sepal.Width
#> 159     setosa 2.9  Sepal.Width
#> 160     setosa 3.1  Sepal.Width
#> 161     setosa 3.7  Sepal.Width
#> 162     setosa 3.4  Sepal.Width
#> 163     setosa 3.0  Sepal.Width
#> 164     setosa 3.0  Sepal.Width
#> 165     setosa 4.0  Sepal.Width
#> 166     setosa 4.4  Sepal.Width
#> 167     setosa 3.9  Sepal.Width
#> 168     setosa 3.5  Sepal.Width
#> 169     setosa 3.8  Sepal.Width
#> 170     setosa 3.8  Sepal.Width
#> 171     setosa 3.4  Sepal.Width
#> 172     setosa 3.7  Sepal.Width
#> 173     setosa 3.6  Sepal.Width
#> 174     setosa 3.3  Sepal.Width
#> 175     setosa 3.4  Sepal.Width
#> 176     setosa 3.0  Sepal.Width
#> 177     setosa 3.4  Sepal.Width
#> 178     setosa 3.5  Sepal.Width
#> 179     setosa 3.4  Sepal.Width
#> 180     setosa 3.2  Sepal.Width
#> 181     setosa 3.1  Sepal.Width
#> 182     setosa 3.4  Sepal.Width
#> 183     setosa 4.1  Sepal.Width
#> 184     setosa 4.2  Sepal.Width
#> 185     setosa 3.1  Sepal.Width
#> 186     setosa 3.2  Sepal.Width
#> 187     setosa 3.5  Sepal.Width
#> 188     setosa 3.6  Sepal.Width
#> 189     setosa 3.0  Sepal.Width
#> 190     setosa 3.4  Sepal.Width
#> 191     setosa 3.5  Sepal.Width
#> 192     setosa 2.3  Sepal.Width
#> 193     setosa 3.2  Sepal.Width
#> 194     setosa 3.5  Sepal.Width
#> 195     setosa 3.8  Sepal.Width
#> 196     setosa 3.0  Sepal.Width
#> 197     setosa 3.8  Sepal.Width
#> 198     setosa 3.2  Sepal.Width
#> 199     setosa 3.7  Sepal.Width
#> 200     setosa 3.3  Sepal.Width
#> 201 versicolor 3.2  Sepal.Width
#> 202 versicolor 3.2  Sepal.Width
#> 203 versicolor 3.1  Sepal.Width
#> 204 versicolor 2.3  Sepal.Width
#> 205 versicolor 2.8  Sepal.Width
#> 206 versicolor 2.8  Sepal.Width
#> 207 versicolor 3.3  Sepal.Width
#> 208 versicolor 2.4  Sepal.Width
#> 209 versicolor 2.9  Sepal.Width
#> 210 versicolor 2.7  Sepal.Width
#> 211 versicolor 2.0  Sepal.Width
#> 212 versicolor 3.0  Sepal.Width
#> 213 versicolor 2.2  Sepal.Width
#> 214 versicolor 2.9  Sepal.Width
#> 215 versicolor 2.9  Sepal.Width
#> 216 versicolor 3.1  Sepal.Width
#> 217 versicolor 3.0  Sepal.Width
#> 218 versicolor 2.7  Sepal.Width
#> 219 versicolor 2.2  Sepal.Width
#> 220 versicolor 2.5  Sepal.Width
#> 221 versicolor 3.2  Sepal.Width
#> 222 versicolor 2.8  Sepal.Width
#> 223 versicolor 2.5  Sepal.Width
#> 224 versicolor 2.8  Sepal.Width
#> 225 versicolor 2.9  Sepal.Width
#> 226 versicolor 3.0  Sepal.Width
#> 227 versicolor 2.8  Sepal.Width
#> 228 versicolor 3.0  Sepal.Width
#> 229 versicolor 2.9  Sepal.Width
#> 230 versicolor 2.6  Sepal.Width
#> 231 versicolor 2.4  Sepal.Width
#> 232 versicolor 2.4  Sepal.Width
#> 233 versicolor 2.7  Sepal.Width
#> 234 versicolor 2.7  Sepal.Width
#> 235 versicolor 3.0  Sepal.Width
#> 236 versicolor 3.4  Sepal.Width
#> 237 versicolor 3.1  Sepal.Width
#> 238 versicolor 2.3  Sepal.Width
#> 239 versicolor 3.0  Sepal.Width
#> 240 versicolor 2.5  Sepal.Width
#> 241 versicolor 2.6  Sepal.Width
#> 242 versicolor 3.0  Sepal.Width
#> 243 versicolor 2.6  Sepal.Width
#> 244 versicolor 2.3  Sepal.Width
#> 245 versicolor 2.7  Sepal.Width
#> 246 versicolor 3.0  Sepal.Width
#> 247 versicolor 2.9  Sepal.Width
#> 248 versicolor 2.9  Sepal.Width
#> 249 versicolor 2.5  Sepal.Width
#> 250 versicolor 2.8  Sepal.Width
#> 251  virginica 3.3  Sepal.Width
#> 252  virginica 2.7  Sepal.Width
#> 253  virginica 3.0  Sepal.Width
#> 254  virginica 2.9  Sepal.Width
#> 255  virginica 3.0  Sepal.Width
#> 256  virginica 3.0  Sepal.Width
#> 257  virginica 2.5  Sepal.Width
#> 258  virginica 2.9  Sepal.Width
#> 259  virginica 2.5  Sepal.Width
#> 260  virginica 3.6  Sepal.Width
#> 261  virginica 3.2  Sepal.Width
#> 262  virginica 2.7  Sepal.Width
#> 263  virginica 3.0  Sepal.Width
#> 264  virginica 2.5  Sepal.Width
#> 265  virginica 2.8  Sepal.Width
#> 266  virginica 3.2  Sepal.Width
#> 267  virginica 3.0  Sepal.Width
#> 268  virginica 3.8  Sepal.Width
#> 269  virginica 2.6  Sepal.Width
#> 270  virginica 2.2  Sepal.Width
#> 271  virginica 3.2  Sepal.Width
#> 272  virginica 2.8  Sepal.Width
#> 273  virginica 2.8  Sepal.Width
#> 274  virginica 2.7  Sepal.Width
#> 275  virginica 3.3  Sepal.Width
#> 276  virginica 3.2  Sepal.Width
#> 277  virginica 2.8  Sepal.Width
#> 278  virginica 3.0  Sepal.Width
#> 279  virginica 2.8  Sepal.Width
#> 280  virginica 3.0  Sepal.Width
#> 281  virginica 2.8  Sepal.Width
#> 282  virginica 3.8  Sepal.Width
#> 283  virginica 2.8  Sepal.Width
#> 284  virginica 2.8  Sepal.Width
#> 285  virginica 2.6  Sepal.Width
#> 286  virginica 3.0  Sepal.Width
#> 287  virginica 3.4  Sepal.Width
#> 288  virginica 3.1  Sepal.Width
#> 289  virginica 3.0  Sepal.Width
#> 290  virginica 3.1  Sepal.Width
#> 291  virginica 3.1  Sepal.Width
#> 292  virginica 3.1  Sepal.Width
#> 293  virginica 2.7  Sepal.Width
#> 294  virginica 3.2  Sepal.Width
#> 295  virginica 3.3  Sepal.Width
#> 296  virginica 3.0  Sepal.Width
#> 297  virginica 2.5  Sepal.Width
#> 298  virginica 3.0  Sepal.Width
#> 299  virginica 3.4  Sepal.Width
#> 300  virginica 3.0  Sepal.Width
#> 301     setosa 1.4 Petal.Length
#> 302     setosa 1.4 Petal.Length
#> 303     setosa 1.3 Petal.Length
#> 304     setosa 1.5 Petal.Length
#> 305     setosa 1.4 Petal.Length
#> 306     setosa 1.7 Petal.Length
#> 307     setosa 1.4 Petal.Length
#> 308     setosa 1.5 Petal.Length
#> 309     setosa 1.4 Petal.Length
#> 310     setosa 1.5 Petal.Length
#> 311     setosa 1.5 Petal.Length
#> 312     setosa 1.6 Petal.Length
#> 313     setosa 1.4 Petal.Length
#> 314     setosa 1.1 Petal.Length
#> 315     setosa 1.2 Petal.Length
#> 316     setosa 1.5 Petal.Length
#> 317     setosa 1.3 Petal.Length
#> 318     setosa 1.4 Petal.Length
#> 319     setosa 1.7 Petal.Length
#> 320     setosa 1.5 Petal.Length
#> 321     setosa 1.7 Petal.Length
#> 322     setosa 1.5 Petal.Length
#> 323     setosa 1.0 Petal.Length
#> 324     setosa 1.7 Petal.Length
#> 325     setosa 1.9 Petal.Length
#> 326     setosa 1.6 Petal.Length
#> 327     setosa 1.6 Petal.Length
#> 328     setosa 1.5 Petal.Length
#> 329     setosa 1.4 Petal.Length
#> 330     setosa 1.6 Petal.Length
#> 331     setosa 1.6 Petal.Length
#> 332     setosa 1.5 Petal.Length
#> 333     setosa 1.5 Petal.Length
#> 334     setosa 1.4 Petal.Length
#> 335     setosa 1.5 Petal.Length
#> 336     setosa 1.2 Petal.Length
#> 337     setosa 1.3 Petal.Length
#> 338     setosa 1.4 Petal.Length
#> 339     setosa 1.3 Petal.Length
#> 340     setosa 1.5 Petal.Length
#> 341     setosa 1.3 Petal.Length
#> 342     setosa 1.3 Petal.Length
#> 343     setosa 1.3 Petal.Length
#> 344     setosa 1.6 Petal.Length
#> 345     setosa 1.9 Petal.Length
#> 346     setosa 1.4 Petal.Length
#> 347     setosa 1.6 Petal.Length
#> 348     setosa 1.4 Petal.Length
#> 349     setosa 1.5 Petal.Length
#> 350     setosa 1.4 Petal.Length
#> 351 versicolor 4.7 Petal.Length
#> 352 versicolor 4.5 Petal.Length
#> 353 versicolor 4.9 Petal.Length
#> 354 versicolor 4.0 Petal.Length
#> 355 versicolor 4.6 Petal.Length
#> 356 versicolor 4.5 Petal.Length
#> 357 versicolor 4.7 Petal.Length
#> 358 versicolor 3.3 Petal.Length
#> 359 versicolor 4.6 Petal.Length
#> 360 versicolor 3.9 Petal.Length
#> 361 versicolor 3.5 Petal.Length
#> 362 versicolor 4.2 Petal.Length
#> 363 versicolor 4.0 Petal.Length
#> 364 versicolor 4.7 Petal.Length
#> 365 versicolor 3.6 Petal.Length
#> 366 versicolor 4.4 Petal.Length
#> 367 versicolor 4.5 Petal.Length
#> 368 versicolor 4.1 Petal.Length
#> 369 versicolor 4.5 Petal.Length
#> 370 versicolor 3.9 Petal.Length
#> 371 versicolor 4.8 Petal.Length
#> 372 versicolor 4.0 Petal.Length
#> 373 versicolor 4.9 Petal.Length
#> 374 versicolor 4.7 Petal.Length
#> 375 versicolor 4.3 Petal.Length
#> 376 versicolor 4.4 Petal.Length
#> 377 versicolor 4.8 Petal.Length
#> 378 versicolor 5.0 Petal.Length
#> 379 versicolor 4.5 Petal.Length
#> 380 versicolor 3.5 Petal.Length
#> 381 versicolor 3.8 Petal.Length
#> 382 versicolor 3.7 Petal.Length
#> 383 versicolor 3.9 Petal.Length
#> 384 versicolor 5.1 Petal.Length
#> 385 versicolor 4.5 Petal.Length
#> 386 versicolor 4.5 Petal.Length
#> 387 versicolor 4.7 Petal.Length
#> 388 versicolor 4.4 Petal.Length
#> 389 versicolor 4.1 Petal.Length
#> 390 versicolor 4.0 Petal.Length
#> 391 versicolor 4.4 Petal.Length
#> 392 versicolor 4.6 Petal.Length
#> 393 versicolor 4.0 Petal.Length
#> 394 versicolor 3.3 Petal.Length
#> 395 versicolor 4.2 Petal.Length
#> 396 versicolor 4.2 Petal.Length
#> 397 versicolor 4.2 Petal.Length
#> 398 versicolor 4.3 Petal.Length
#> 399 versicolor 3.0 Petal.Length
#> 400 versicolor 4.1 Petal.Length
#> 401  virginica 6.0 Petal.Length
#> 402  virginica 5.1 Petal.Length
#> 403  virginica 5.9 Petal.Length
#> 404  virginica 5.6 Petal.Length
#> 405  virginica 5.8 Petal.Length
#> 406  virginica 6.6 Petal.Length
#> 407  virginica 4.5 Petal.Length
#> 408  virginica 6.3 Petal.Length
#> 409  virginica 5.8 Petal.Length
#> 410  virginica 6.1 Petal.Length
#> 411  virginica 5.1 Petal.Length
#> 412  virginica 5.3 Petal.Length
#> 413  virginica 5.5 Petal.Length
#> 414  virginica 5.0 Petal.Length
#> 415  virginica 5.1 Petal.Length
#> 416  virginica 5.3 Petal.Length
#> 417  virginica 5.5 Petal.Length
#> 418  virginica 6.7 Petal.Length
#> 419  virginica 6.9 Petal.Length
#> 420  virginica 5.0 Petal.Length
#> 421  virginica 5.7 Petal.Length
#> 422  virginica 4.9 Petal.Length
#> 423  virginica 6.7 Petal.Length
#> 424  virginica 4.9 Petal.Length
#> 425  virginica 5.7 Petal.Length
#> 426  virginica 6.0 Petal.Length
#> 427  virginica 4.8 Petal.Length
#> 428  virginica 4.9 Petal.Length
#> 429  virginica 5.6 Petal.Length
#> 430  virginica 5.8 Petal.Length
#> 431  virginica 6.1 Petal.Length
#> 432  virginica 6.4 Petal.Length
#> 433  virginica 5.6 Petal.Length
#> 434  virginica 5.1 Petal.Length
#> 435  virginica 5.6 Petal.Length
#> 436  virginica 6.1 Petal.Length
#> 437  virginica 5.6 Petal.Length
#> 438  virginica 5.5 Petal.Length
#> 439  virginica 4.8 Petal.Length
#> 440  virginica 5.4 Petal.Length
#> 441  virginica 5.6 Petal.Length
#> 442  virginica 5.1 Petal.Length
#> 443  virginica 5.1 Petal.Length
#> 444  virginica 5.9 Petal.Length
#> 445  virginica 5.7 Petal.Length
#> 446  virginica 5.2 Petal.Length
#> 447  virginica 5.0 Petal.Length
#> 448  virginica 5.2 Petal.Length
#> 449  virginica 5.4 Petal.Length
#> 450  virginica 5.1 Petal.Length
#> 451     setosa 0.2  Petal.Width
#> 452     setosa 0.2  Petal.Width
#> 453     setosa 0.2  Petal.Width
#> 454     setosa 0.2  Petal.Width
#> 455     setosa 0.2  Petal.Width
#> 456     setosa 0.4  Petal.Width
#> 457     setosa 0.3  Petal.Width
#> 458     setosa 0.2  Petal.Width
#> 459     setosa 0.2  Petal.Width
#> 460     setosa 0.1  Petal.Width
#> 461     setosa 0.2  Petal.Width
#> 462     setosa 0.2  Petal.Width
#> 463     setosa 0.1  Petal.Width
#> 464     setosa 0.1  Petal.Width
#> 465     setosa 0.2  Petal.Width
#> 466     setosa 0.4  Petal.Width
#> 467     setosa 0.4  Petal.Width
#> 468     setosa 0.3  Petal.Width
#> 469     setosa 0.3  Petal.Width
#> 470     setosa 0.3  Petal.Width
#> 471     setosa 0.2  Petal.Width
#> 472     setosa 0.4  Petal.Width
#> 473     setosa 0.2  Petal.Width
#> 474     setosa 0.5  Petal.Width
#> 475     setosa 0.2  Petal.Width
#> 476     setosa 0.2  Petal.Width
#> 477     setosa 0.4  Petal.Width
#> 478     setosa 0.2  Petal.Width
#> 479     setosa 0.2  Petal.Width
#> 480     setosa 0.2  Petal.Width
#> 481     setosa 0.2  Petal.Width
#> 482     setosa 0.4  Petal.Width
#> 483     setosa 0.1  Petal.Width
#> 484     setosa 0.2  Petal.Width
#> 485     setosa 0.2  Petal.Width
#> 486     setosa 0.2  Petal.Width
#> 487     setosa 0.2  Petal.Width
#> 488     setosa 0.1  Petal.Width
#> 489     setosa 0.2  Petal.Width
#> 490     setosa 0.2  Petal.Width
#> 491     setosa 0.3  Petal.Width
#> 492     setosa 0.3  Petal.Width
#> 493     setosa 0.2  Petal.Width
#> 494     setosa 0.6  Petal.Width
#> 495     setosa 0.4  Petal.Width
#> 496     setosa 0.3  Petal.Width
#> 497     setosa 0.2  Petal.Width
#> 498     setosa 0.2  Petal.Width
#> 499     setosa 0.2  Petal.Width
#> 500     setosa 0.2  Petal.Width
#> 501 versicolor 1.4  Petal.Width
#> 502 versicolor 1.5  Petal.Width
#> 503 versicolor 1.5  Petal.Width
#> 504 versicolor 1.3  Petal.Width
#> 505 versicolor 1.5  Petal.Width
#> 506 versicolor 1.3  Petal.Width
#> 507 versicolor 1.6  Petal.Width
#> 508 versicolor 1.0  Petal.Width
#> 509 versicolor 1.3  Petal.Width
#> 510 versicolor 1.4  Petal.Width
#> 511 versicolor 1.0  Petal.Width
#> 512 versicolor 1.5  Petal.Width
#> 513 versicolor 1.0  Petal.Width
#> 514 versicolor 1.4  Petal.Width
#> 515 versicolor 1.3  Petal.Width
#> 516 versicolor 1.4  Petal.Width
#> 517 versicolor 1.5  Petal.Width
#> 518 versicolor 1.0  Petal.Width
#> 519 versicolor 1.5  Petal.Width
#> 520 versicolor 1.1  Petal.Width
#> 521 versicolor 1.8  Petal.Width
#> 522 versicolor 1.3  Petal.Width
#> 523 versicolor 1.5  Petal.Width
#> 524 versicolor 1.2  Petal.Width
#> 525 versicolor 1.3  Petal.Width
#> 526 versicolor 1.4  Petal.Width
#> 527 versicolor 1.4  Petal.Width
#> 528 versicolor 1.7  Petal.Width
#> 529 versicolor 1.5  Petal.Width
#> 530 versicolor 1.0  Petal.Width
#> 531 versicolor 1.1  Petal.Width
#> 532 versicolor 1.0  Petal.Width
#> 533 versicolor 1.2  Petal.Width
#> 534 versicolor 1.6  Petal.Width
#> 535 versicolor 1.5  Petal.Width
#> 536 versicolor 1.6  Petal.Width
#> 537 versicolor 1.5  Petal.Width
#> 538 versicolor 1.3  Petal.Width
#> 539 versicolor 1.3  Petal.Width
#> 540 versicolor 1.3  Petal.Width
#> 541 versicolor 1.2  Petal.Width
#> 542 versicolor 1.4  Petal.Width
#> 543 versicolor 1.2  Petal.Width
#> 544 versicolor 1.0  Petal.Width
#> 545 versicolor 1.3  Petal.Width
#> 546 versicolor 1.2  Petal.Width
#> 547 versicolor 1.3  Petal.Width
#> 548 versicolor 1.3  Petal.Width
#> 549 versicolor 1.1  Petal.Width
#> 550 versicolor 1.3  Petal.Width
#> 551  virginica 2.5  Petal.Width
#> 552  virginica 1.9  Petal.Width
#> 553  virginica 2.1  Petal.Width
#> 554  virginica 1.8  Petal.Width
#> 555  virginica 2.2  Petal.Width
#> 556  virginica 2.1  Petal.Width
#> 557  virginica 1.7  Petal.Width
#> 558  virginica 1.8  Petal.Width
#> 559  virginica 1.8  Petal.Width
#> 560  virginica 2.5  Petal.Width
#> 561  virginica 2.0  Petal.Width
#> 562  virginica 1.9  Petal.Width
#> 563  virginica 2.1  Petal.Width
#> 564  virginica 2.0  Petal.Width
#> 565  virginica 2.4  Petal.Width
#> 566  virginica 2.3  Petal.Width
#> 567  virginica 1.8  Petal.Width
#> 568  virginica 2.2  Petal.Width
#> 569  virginica 2.3  Petal.Width
#> 570  virginica 1.5  Petal.Width
#> 571  virginica 2.3  Petal.Width
#> 572  virginica 2.0  Petal.Width
#> 573  virginica 2.0  Petal.Width
#> 574  virginica 1.8  Petal.Width
#> 575  virginica 2.1  Petal.Width
#> 576  virginica 1.8  Petal.Width
#> 577  virginica 1.8  Petal.Width
#> 578  virginica 1.8  Petal.Width
#> 579  virginica 2.1  Petal.Width
#> 580  virginica 1.6  Petal.Width
#> 581  virginica 1.9  Petal.Width
#> 582  virginica 2.0  Petal.Width
#> 583  virginica 2.2  Petal.Width
#> 584  virginica 1.5  Petal.Width
#> 585  virginica 1.4  Petal.Width
#> 586  virginica 2.3  Petal.Width
#> 587  virginica 2.4  Petal.Width
#> 588  virginica 1.8  Petal.Width
#> 589  virginica 1.8  Petal.Width
#> 590  virginica 2.1  Petal.Width
#> 591  virginica 2.4  Petal.Width
#> 592  virginica 2.3  Petal.Width
#> 593  virginica 1.9  Petal.Width
#> 594  virginica 2.3  Petal.Width
#> 595  virginica 2.5  Petal.Width
#> 596  virginica 2.3  Petal.Width
#> 597  virginica 1.9  Petal.Width
#> 598  virginica 2.0  Petal.Width
#> 599  virginica 2.3  Petal.Width
#> 600  virginica 1.8  Petal.Width