Compose a stable isotope box model scenario

A function to compose an isobxr box model scenario, defined by a series of n successive runs,
each run inheriting from the final state conditions of the previous run.
It is possible to force parameters at each run, namely:

fluxes
(overwriting all or a subset of fluxes defined in 0_ISOBXR_MASTER.xlsx master file)

isotope fractionation coefficients
(overwriting all or a subset of coefficients defined in 0_ISOBXR_MASTER.xlsx master file)

box sizes
(overwriting all or a subset of box sizes defined in 0_ISOBXR_MASTER.xlsx master file)

rayleigh isotope distillation

isotope composition of a source box at initial state

compose_isobxr(
  workdir,
  SERIES_ID,
  time_units,
  COMPO_MASTER,
  plot_HIDE_BOXES_delta = NULL,
  plot_HIDE_BOXES_size = NULL,
  EACH_RUN_DIGEST = FALSE,
  to_CPS_DIGEST_CSVs = FALSE
)

Arguments

workdir
Working directory of 0_ISOBXR_MASTER.xlsx master file,
of the composite master file (e.g., 0_COMPO_MASTER.xlsx)
and where output files will be stored.
(character string)

SERIES_ID
Name of the composite model series the run belongs to.
It determines the folder in which the output files will be stored for this composite run.
A composite run number is automatically linked to it, subsequent runs can not overwrite a previous composite run.
(character string)

time_units
Vector defining the initial time unit (identical to unit used in fluxes),
followed by the time unit used for the graphical output.
Character string, to be selected amongst the following:
micros, ms, s, min, h, d, wk, mo, yr, kyr, Myr, Gyr
e.g., c("d", "yr") to convert days into years

COMPO_MASTER
Name of the composite master file (e.g., 0_COMPO_MASTER.xlsx), defining the composite run scenario.
(character string)

plot_HIDE_BOXES_delta
OPTIONAL
Vector of character strings,
defining the names of the boxes to hide in the plot of the delta values as a function of time, edited as a pdf.
(e.g., c("BOX_A", "BOX_C"))
Default is NULL (no box hidden).

plot_HIDE_BOXES_size
OPTIONAL
Vector of character strings,
defining the names of the boxes to hide in the plot of the box sizes (masses of X) as a function of time, edited as a pdf.
(e.g., c("BOX_A", "BOX_C"))
Default is NULL (no box hidden).

EACH_RUN_DIGEST
OPTIONAL
Logical value.
Edits full digests for each model run (all optional outputs of run_isobxr function) if TRUE.
Default is FALSE.

to_CPS_DIGEST_CSVs
OPTIONAL
Logical value.
Exports all global csv outputs to 0_CPS_DIGEST folder (full evD and full evS) if TRUE.
Default is FALSE.

Value

Creates and stores all outputs in a dedicated composite SERIES directory located in working directory, with the following name structure:
3_CPS + SERIES_ID + YYY, where YYY is a composite scenario number automically set between 001 and 999.
No overwriting of previous composite runs is possible.

Creates the set of inputs and outputs for all successive n runs,
numbered from to 1 to n in an XXXX format with the following format:
CPS + SERIES_ID + YYY + XXXX + IN.Rda
CPS + SERIES_ID + YYY + XXXX + OUT.Rda
(see run_isobxr documentation)

Writes summarized results in the 0_CPS_DIGEST folder:

LOG file of local composite run.
(file name structure: CPS + SERIES_ID + YYY + _LOG.csv)

Composite master file.
(file name structure: CPS + SERIES_ID + YYY + _MASTER.xlsx)

Dataset of temporal evolution of delta values (evD) in all boxes over the n runs that constitute the composite run scenario
(file name structure: CPS + SERIES_ID + YYY + evD.RDS)

Dataset of temporal evolution of box sizes (evS, masses of X) in all boxes over the n runs that constitute the composite run scenario
(file name structure: CPS + SERIES_ID + YYY + evS.RDS)

All-in-one plot of the evolution of delta values + sizes in all non hidden boxes.
(file name structure: CPS + SERIES_ID + YYY + p_evDS.pdf)

Multiple plots of the evolution of delta values in all non hidden boxes.
(file name structure: CPS + SERIES_ID + YYY + pf_evD.pdf)

Multiple plots of the evolution of box sizes in all non hidden boxes.
(file name structure: CPS + SERIES_ID + YYY + pf_evS.pdf)

Optional outputs

If EACH_RUN_DIGEST = TRUE
Creates and fills DIGEST folder for each run of the composite scenario with all optional outputs of run_isobxr function.
(folder name structure: CPS + SERIES_ID + YYY + XXXX + DIGEST)

If to_CPS_DIGEST_CSVs = TRUE
In the 0_CPS_DIGEST folder, edits csv versions of the whole-composite scenario evD and evS datasets.
(file names structures: CPS + SERIES_ID + YYY + evD.csv and CPS + SERIES_ID + YYY + evS.csv)

See also