Numerically solve stable isotope box models

A numerical solver of the system of ordinary differential equations (ODES), describing the evolution of stable isotope ratios in all boxes of a system.
Not intended for direct use although possible.
The numerical solver uses the ode function of the deSolve package to integrate the stable isotopes ratios over time in each box. It allows the calculation of the evolution of stable isotope ratio even in the case of unbalanced outward and inward fluxes of element X in a given box resulting in the accumulation or loss of element X.

num_slvr(input_path, to_DIGEST_csv = FALSE, save_run_outputs = FALSE)

Arguments

input_path
path to the INPUT file containing all commands for the run
(character string, file name structure: RUN name + _IN.Rda)

to_DIGEST_csv
Edits csv outputs to the RUN DIGEST folder
(logical, default is FALSE)

save_run_outputs
OPTIONAL
Logical value.
Allows saving all run outputs to directory containing INPUT file.
By default, run outputs are stored in a temporary directory and are erased if not exported.
Default is FALSE.

Value

The function returns the numerically determined evolution of stable isotope compositions and mass of element X in all boxes over the run duration as specified in INPUT file.

By default (unless save_run_outputs = TRUE), run outputs are stored in the temporary directory and are not exported.

The outputs of the run are stored in a Rda output file with the following file name structure:
RUN name + _OUT.Rda

Optional csv outputs to the DIGEST folder are as follows

OUT data file storing initial and final size and delta values in all boxes.
(file name structure: out_1_N_OUT + RUN name + .csv)

evS data file storing the evolution with time of the sizes (masses of element X) of all boxes.
(file name structure: out_2_N_evS + RUN name + .csv)

evD data file storing the evolution with time of the delta values in all boxes.
(file name structure: out_3_N_evD + RUN name + .csv)