Run isobxr stable isotope box model

A function to run the isobxr stable isotope box model, assessing the design of the model and automatically running num_slvr or ana_slvr depending on the conditions.

run_isobxr(
  workdir,
  SERIES_ID,
  flux_list_name,
  coeff_list_name,
  t_lim,
  nb_steps,
  time_units,
  FORCING_RAYLEIGH = NULL,
  FORCING_SIZE = NULL,
  FORCING_DELTA = NULL,
  FORCING_ALPHA = NULL,
  COMPOSITE = FALSE,
  COMPO_SERIES_n = NaN,
  COMPO_SERIES_FAMILY = NaN,
  EXPLORER = FALSE,
  EXPLO_SERIES_n = NaN,
  EXPLO_SERIES_FAMILY = NaN,
  HIDE_PRINTS = FALSE,
  PLOT_DIAGRAMS = TRUE,
  PLOT_evD = FALSE
)

Arguments

workdir
Working directory in which the master file (0_ISOBXR_MASTER.xlsx) is found and where the output files will be stored.
(string of characters)

SERIES_ID
Name of the model series the run belongs to, that will determine the folder in which the output files will be stored.
(string of characters)

flux_list_name
Name of the list of fluxes to be used for the run as defined in a single column of the FLUXES sheet of the 0_ISOBXR_MASTER.xlsx file.
(string of characters)

coeff_list_name
Name of the list of fractionation coefficients to be used for the run as defined in a single column of the COEFFS sheet of the 0_ISOBXR_MASTER.xlsx file.
(string of characters)

t_lim
Run duration, given in the same time units as the fluxes.
(integer)

nb_steps
Number of steps that will define the resolution of the run.
(integer)

time_units
Vector defining the initial time units (identical to time unit used for fluxes) followed by the time unit used for the graphic output.
(vector of two strings of characters, eg. c("days", "years"))

FORCING_RAYLEIGH
OPTIONAL Dataframe describing the forcing on a Rayleigh distillation driven fractionation coefficient as a function of flux intensities and a fundamental fractionation coefficient (Dataframe formating details in read me).
Default is NULL.

FORCING_SIZE
OPTIONAL Dataframe describing the forcing on one or several box sizes (mass of element X). The newly defined sizes for the given set of boxes shown in dataframe will overwrite the sizes as defined in 0_ISOBXR_MASTER.xlsx file.
Default is NULL.

FORCING_DELTA
OPTIONAL Dataframe describing the forcing on one or several boxes initial isotope composition described as delta (Dataframe formating details in read me). The newly defined delta values for the given set of boxes shown in dataframe will overwrite the delta values as defined in ISOBXR_MASTER file.
Default is NULL.

FORCING_ALPHA
OPTIONAL Dataframe describing the forcing on one or several fractionation coefficients linked to fluxes from one reservoir to another (Dataframe formating details in read me). The newly defined alpha values for the given set of boxes shown in dataframe will overwrite the alpha values as defined in ISOBXR_MASTER file.
Default is NULL.

COMPOSITE
NOT TO BE USED IN SINGLE RUN Logical value automatically defined in compose_isobxr.
Default is FALSE.

COMPO_SERIES_n
NOT TO BE USED IN SINGLE RUN Iteration of the composite run for the given series it belongs to, automatically defined in compose_isobxr.
Default is NaN.

COMPO_SERIES_FAMILY
NOT TO BE USED IN SINGLE RUN Composite run series family, automatically defined in compose_isobxr.
Default is NaN.

EXPLORER
NOT TO BE USED IN SINGLE RUN Logical value automatically defined in sweep_steady or sweep_dyn.
Default is FALSE.

EXPLO_SERIES_n
NOT TO BE USED IN SINGLE RUN Iteration of the sweep run for the given series it belongs to, automatically defined in sweep_steady or sweep_dyn.
Default is NaN.

EXPLO_SERIES_FAMILY
NOT TO BE USED IN SINGLE RUN Sweep run series family, automatically defined in sweep_steady or sweep_dyn.
Default is NaN.

HIDE_PRINTS
OPTIONAL Logical value determining whether to run outputs details in R console. This parameter will not hide the warnings regarding the automatic update of the run duration in case of the emptying of a box.
Default is FALSE.

PLOT_DIAGRAMS
OPTIONAL Logical value to edit pdf of box model diagram or not.
Default is TRUE.

PLOT_evD
OPTIONAL Logical value to edit pdf of delta time evolution plot or not.
Default is TRUE.

Value

If function is run independently and if the directory is non existing, the fonction creates and stores all outputs in a SERIES directory located in working directory.
Directory name structure: 2_RUN + SERIES_ID

Automatically sets a XXXX run number between 0001 and 9999. The outputs do not overwrite possible identical previously performed runs.

Creates an INPUT file with all run conditions, located in SERIES directory.
(file name structure: SERIES_ID + XXXX + _INPUT.xlsx)

Edits a Box model diagram of flux (DIAGf pdf) of element X (mass per time unit) between all boxes. (Optional, depends on parameter PLOT_DIAGRAM), located in SERIES directory.
(file name structure: SERIES_ID + XXXX + DIAGf + _flux_list_name.pdf)

Edits a Box model diagram of isotope fractionation coefficients (alphas, DIAGa pdf) between all boxes (optional, depends on parameter PLOT_DIAGRAM), located in SERIES directory.
(file name structure: SERIES_ID + XXXX + DIAGf + _flux_list_name.pdf)

Creates a subdirectory for numerical outputs.
(directory name structure: SERIES_ID + XXXX + _OUT)

Stores num_slvr or ana_slvr outputs in subdirectory.

Edits a pdf plot of the time dependent evolution of delta values, with a logarithmic x axis time scale (optional).
(file name structure: SERIES_ID + XXXX + _plot_evD.pdf)

Creates or updates the general log file, located in general working directory.
(file name: 1_LOG.csv)

See also