Analytically solve stable isotope box models

An analytical solver of the system of ordinary differential
equations (ODES) of stable isotope ratios of element X in all boxes.
Not intended for manual use.
The analytical solver finds the eigenvalues and eigenvectors of the ODES.
It determines the set of analytical solutions that describes the evolution of isotope ratios in each box over time.

solve_analytically(IN, paths, to_DIGEST_csv = FALSE, return_results = FALSE)

Arguments

IN

input data, edited by sim.single_run (list of dataframes)

paths

paths object edited by sim.single_run (list of characters)

to_DIGEST_csv

if TRUE, edits csv outputs to DIGEST directory
Default is FALSE.

return_results

if TRUE, results returned as a list of R objects.
Default is FALSE.

Value

Analytically determined evolution of stable isotope compositions in all boxes over the run duration as specified in INPUT file.

Run outputs are stored in a temporary directory and not exported by solve_numerically.

The outputs of the run are stored in the rds output file in the SERIES directory with the following file name structure:

SERIES_ID + RUN_n.rds

Optional csv outputs to the DIGEST folder are as follows

OUT data file with initial and final size and delta values in all boxes.
(file name structure: out_1_A_OUT + SERIES_ID + RUN_n + .csv)

ODE_SOLNs data file summarizing outputs of the analytical solutions of the ODES
(eigenvalues, eigenvectors, relaxation times, constants according to initial conditions).
(file name structure: out_2_A_ODE_SOLNs + SERIES_ID + RUN_n + .csv)

evD data file of the evolution with time of the delta values in all boxes.
(file name structure: out_3_A_evD + SERIES_ID + RUN_n + .csv)