Sweep the space of n parameters at the final state of a system

A function to assess the influence of n parameters (varying over a range of values) on the final state of a system.

sweep.final_nD(
  workdir,
  sweep_master_file,
  sweep_dir_to_complete = NULL,
  export.data_as_csv_xlsx = FALSE,
  isobxr_master_file = "0_ISOBXR_MASTER",
  save_outputs = FALSE
)

Arguments

workdir

Working directory of 0_ISOBXR_MASTER.xlsx master file,
of the dynamic sweep master file (e.g., 0_EXPLO_DYN_MASTER.xlsx)
and where output files will be stored if saved by user.
(character string)

sweep_master_file

Name of sweep.final_nD excel master file. (without file "xlsx" extension).

sweep_dir_to_complete

Name of directory of previously halted sweep.final_nD run that the user wishes to continue. Starts with "4_FINnD".
Default is NULL.

export.data_as_csv_xlsx

If TRUE, exports chunk sweep result data as csv and xlsx fo full to chunk digest directory.
Default is FALSE.

isobxr_master_file

Name of isobxr excel master file.
Default is "0_ISOBXR_MASTER".

save_outputs

If TRUE, saves all run outputs to local working directory (workdir).
By default, run outputs are stored in a temporary directory and erased if not saved. Default is FALSE.

Value

Delta values and box sizes at final state of the system, in the n-dimensions space of parameters. sweep.final_nD outputs are saved to workdir if save_outputs = TRUE.

sweep.final_nD outputs

digest sweep.final_nD outputs full sweep.final_nD outputs are stored in a digest directory directly in workdir:

isobxr master file archive as xlsx

global chunk_log as csv

global chunk_log as rds

sweep.final_nD master file archive as xlsx

sweep.final_nD merged chunks LOG excerpt as csv

sweep.final_nD merged param_space, current progress, as rds (space of swept parameters)

sweep.final_nD merged results, current progress, as rds

data frame containing delta and size at final state across whole n-dimensions space of parameters

sweep.final_nD param_space as planned, as rds

sweep.final_nD sweep_default data frame of default run conditions, as rds

sweep.final_nD sweep progress plot as pdf, follows merging of chunks

chunks sweep.final_nD outputs chunk directories are edited in case total number. of runs exceeds chunk size defined in sweep.final_nD master file. They are removed after full sweep.