Sweep the space of two parameters during a dynamic run

A function to assess the influence of two parameters (varying over a range of values) on dynamic evolution of given model.

sweep_dyn(
  workdir,
  SERIES_ID,
  time_units,
  EXPLO_MASTER,
  EXPLO_AXIS_1,
  EXPLO_AXIS_2
)

Arguments

workdir
Working directory in which the general master file (0_ISOBXR_MASTER.xlsx) and dynamic sweep master file (e.g., 0_EXPLO_DYN_MASTER.xlsx) are found and where the output files will be stored.
(string of characters)

SERIES_ID
Name of the sweep series the run belongs to, that will determine the folder in which the output files of this sweep run will be stored. A sweep run number in YYY format will be automatically linked to it and no run will overwrite a previous sweep run.
(string of characters)

time_units
Vector defining the initial time units (identical to time unit used for fluxes) followed by the time unit used for the graphic output.
(vector of two strings of characters, eg. c("days", "years"))

EXPLO_MASTER
Name of the dynamic sweep master file (e.g., 0_EXPLO_DYN_MASTER.xlsx), defining the dynamic sweep run scenario.
(string of characters)

EXPLO_AXIS_1
Set of values of sweeping parameter 1

EXPLO_AXIS_2
Set of values of sweeping parameter 2

Value

If non existing, the fonction creates and stores all outputs in a SERIES directory located in working directory, automatically numbered.
Directory name structure: 4_DYN + SERIES_ID + YYY, YYY being an automically set dynamic sweep run number between 001 and 999.
No overwriting of previous dynamic sweep runs is possible.

Calculates the number of runs the sweeping will require depending on the sweeped parameters. The function then asks the user confirmation to run sweep_dyn, as the run calculation time depends on the number of successive sweeping runs.

Creates the set of inputs and outputs for each successive n runs, numbered from to 1 to n in an XXXX format (see run_isobxr documentation).
Single run plots normally edited by run_isobxr are here not edited.
Named with the following format: DYN + SERIES_ID + YYY + XXXX

Archives the dynamic sweep master file.
(file name structure: 0_DYN + SERIES_ID + YYY + 0_EXPLO_DYN_MASTER.xlsx)

Archives the local LOG for the given dynamic sweep run.
(file name structure: 0_DYN + SERIES_ID + YYY + LOG.csv)

Stores the evolution of delta values with time in all boxes over the n runs that constitute the dynamic sweep run.
(file name structure: 0_DYN + SERIES_ID + YYY + evD.csv)

Stores the evolution of box sizes (masses of element X) with time in all boxes over the n runs that constitute the dynamic sweep run
(file name structure: 0_DYN + SERIES_ID + YYY + evS.csv)

Stores the final state of delta values with time in all boxes over the n runs that constitute the dynamic sweep run.
(file name structure: 0_DYN + SERIES_ID + YYY + evD_final.csv)

Stores the final state of box sizes (masses of element X) with time in all boxes over the n runs that constitute the dynamic sweep run
(file name structure: 0_DYN + SERIES_ID + YYY + evS_final.csv)