Sweep the space of two parameters at the final state of a run

A function to assess the influence of two parameters (varying over a range of values) on the final state of a given model.

sweep_steady(
  workdir,
  SERIES_ID,
  time_units,
  EXPLO_MASTER,
  EXPLO_AXIS_1,
  EXPLO_AXIS_2,
  to_STD_DIGEST_CSVs = FALSE
)

Arguments

workdir
Working directory of 0_ISOBXR_MASTER.xlsx master file,
of the steady sweep master file (e.g., 0_EXPLO_STEADY_MASTER.xlsx)
and where output files will be stored.
(character string)

SERIES_ID
Name of the sweep series the run belongs to.
It determines the folder in which the output files will be stored for this sweep run.
A sweep run number is automatically linked to it, and subsequent sweep runs can not overwrite a previous one.
(character string)

time_units
Vector defining the initial time unit (identical to unit used in fluxes),
followed by the time unit used for the graphical output.
Character string, to be selected amongst the following:
micros, ms, s, min, h, d, wk, mo, yr, kyr, Myr, Gyr
e.g., c("d", "yr") to convert days into years

EXPLO_MASTER
Name of the steady sweep master file (e.g., 0_EXPLO_STEADY_MASTER.xlsx), defining the steady sweep run scenario.
(character string)

EXPLO_AXIS_1
Set of values of sweeping parameter 1.
See Vignette for further details.

EXPLO_AXIS_2
Set of values of sweeping parameter 2.
See Vignette for further details.

to_STD_DIGEST_CSVs
OPTIONAL
Logical value.
Exports all global csv outputs to 0_STD_DIGEST folder (full and final evD and evS) if TRUE.
Default is FALSE.

Value

Creates and stores all outputs in a dedicated sweep steady SERIES directory located in working directory, with the following name structure:
4_STD + SERIES_ID + YYY, where YYY is a sweep steady run number automically set between 001 and 999.
No overwriting of previous sweep steady run runs is possible.

Calculates the number of single runs the sweeping will require depending on the sweeped parameters.

Asks the user confirmation to run sweep_steady, as the run calculation time depends on the number of successive sweeping runs.

Writes the set of inputs and outputs for the single initial run only with the following format:
STD + SERIES_ID + YYY + 0001 + IN.Rda
STD + SERIES_ID + YYY + 0001 + OUT.Rda
(see run_isobxr documentation).

Writes summarized results in the 0_STD_DIGEST folder:

Archived LOG file of local sweep steady run.
(file name structure: STD + SERIES_ID + YYY + _LOG.csv)

Archived sweep steady master file.
(file name structure: STD + SERIES_ID + YYY + _MASTER.xlsx)

Dataset of temporal evolution of delta values (evD) in all boxes over the n runs that constitute the sweep steady run.
(file name structure: STD + SERIES_ID + YYY + evD.RDS)

Dataset of temporal evolution of box sizes (evS, masses of X) in all boxes over the n runs that constitute the sweep steady run.
(file name structure: STD + SERIES_ID + YYY + evS.RDS)

Dataset of final state of delta values (evD) in all boxes over the n runs that constitute the sweep steady run.
(file name structure: STD + SERIES_ID + YYY + evD_final.RDS)

Dataset of final state of box sizes (evS, masses of X) in all boxes over the n runs that constitute the sweep steady run.
(file name structure: STD + SERIES_ID + YYY + evS_final.RDS)

Optional output

If to_STD_DIGEST_CSVs = TRUE
In the 0_STD_DIGEST folder, edits csv versions of the sweep steady run datasets (full and final evD and evS) with the following name structures:

STD + SERIES_ID + YYY + evD.csv

STD + SERIES_ID + YYY + evS.csv

STD + SERIES_ID + YYY + evD_final.csv

STD + SERIES_ID + YYY + evS_final.csv