Compose a stable isotope box model scenario

A function to compose an isobxr box model scenario, defined by a series of n successive runs,
each run inheriting from the final state conditions of the previous run.
It is possible to force parameters at each run, namely:

fluxes
(overwriting all or a subset of fluxes defined in 0_ISOBXR_MASTER.xlsx master file)

isotope fractionation coefficients
(overwriting all or a subset of coefficients defined in 0_ISOBXR_MASTER.xlsx master file)

box sizes
(overwriting all or a subset of box sizes defined in 0_ISOBXR_MASTER.xlsx master file)

rayleigh isotope distillation

isotope composition of a source box at initial state

compose_isobxr(
  workdir,
  SERIES_ID,
  time_units,
  COMPO_MASTER,
  plot_HIDE_BOXES_delta = NULL,
  plot_HIDE_BOXES_size = NULL,
  EACH_RUN_DIGEST = FALSE,
  to_CPS_DIGEST_CSVs = FALSE,
  plot_results = TRUE,
  save_run_outputs = FALSE
)

Arguments

workdir
Working directory of 0_ISOBXR_MASTER.xlsx master file,
of the composite master file (e.g., 0_COMPO_MASTER.xlsx)
and where output files will be stored if exported by user.
(character string)

SERIES_ID
Name of the composite model series the run belongs to.
It determines the folder in which the output files will be stored for this composite run.
A composite run number is automatically linked to it, subsequent runs can not overwrite a previous composite run.
(character string)

time_units
Vector defining the initial time unit (identical to unit used in fluxes),
followed by the time unit used for the graphical output.
Character string, to be selected among the following:
micros, ms, s, min, h, d, wk, mo, yr, kyr, Myr, Gyr
e.g., c("d", "yr") to convert days into years

COMPO_MASTER
Name of the composite master file (e.g., 0_COMPO_MASTER.xlsx), defining the composite run scenario.
(character string)

plot_HIDE_BOXES_delta
OPTIONAL
Vector of character strings,
defining the names of the boxes to hide in the plot of the delta values as a function of time, edited as a pdf.
(e.g., c("BOX_A", "BOX_C"))
Default is NULL (no box hidden).

plot_HIDE_BOXES_size
OPTIONAL
Vector of character strings,
defining the names of the boxes to hide in the plot of the box sizes (masses of X) as a function of time, edited as a pdf.
(e.g., c("BOX_A", "BOX_C"))
Default is NULL (no box hidden).

EACH_RUN_DIGEST
OPTIONAL
Logical value.
Edits full digests for each model run (all optional outputs of run_isobxr function) if TRUE.
Default is FALSE.

to_CPS_DIGEST_CSVs
OPTIONAL
Logical value.
Exports all global csv outputs to 0_CPS_DIGEST folder (full evD and full evS) if TRUE.
Default is FALSE.

plot_results
OPTIONAL
Logical value.
If TRUE, plots in R session the composite model run evolution of delta values and box sizes for boxes of interest (see plot_HIDE_BOXES_delta and plot_HIDE_BOXES_size parameters to remove boxes from plots).
Default is TRUE.

save_run_outputs
OPTIONAL
Logical value.
Allows saving all run outputs to working directory (workdir).
By default, run outputs are stored in the temporary directory and are erased if not exported.
Default is FALSE.

Value

Calculates the time evolution of delta values and box sizes in all boxes throughout scenario.

compose_isobxr returns by default a plot showing time evolution of delta values and box sizes for all boxes
(set plot_deltas = FALSE to mute the plots)

The graphical results of the composite run can be also interactively explored using the shinobxr_app function in case user saves the outputs to the working directory (save_run_outputs = TRUE).

compose_isobxr creates a series of isotope data and metadata, all of which are stored in a temporary directory.
The user can save all outputs described below to their working directory by setting save_run_outputs = TRUE (default is FALSE).

compose_isobxr creates and stores all outputs in a dedicated dynamic steady SERIES directory with the following name structure:
3_CPS + SERIES_ID + YYY, where YYY is a composite scenario number automatically set between 001 and 999.
No overwriting of previous composite runs is possible.

compose_isobxr base workflow:

Creates the set of inputs and outputs for all successive n runs,
numbered from to 1 to n in an XXXX format with the following format:
CPS + SERIES_ID + YYY + XXXX + IN.Rda
CPS + SERIES_ID + YYY + XXXX + OUT.Rda
(see run_isobxr documentation)

Writes summarized results in the 0_CPS_DIGEST folder:

LOG file of local composite run.
(file name structure: CPS + SERIES_ID + YYY + _LOG.csv)

Composite master file.
(file name structure: CPS + SERIES_ID + YYY + _MASTER.xlsx)

Dataset of temporal evolution of delta values (evD) in all boxes over the n runs that constitute the composite run scenario
(file name structure: CPS + SERIES_ID + YYY + evD.RDS)

Dataset of temporal evolution of box sizes (evS, masses of X) in all boxes over the n runs that constitute the composite run scenario
(file name structure: CPS + SERIES_ID + YYY + evS.RDS)

All-in-one plot of the evolution of delta values + sizes in all non hidden boxes.
(file name structure: CPS + SERIES_ID + YYY + p_evDS.pdf)

Multiple plots of the evolution of delta values in all non hidden boxes.
(file name structure: CPS + SERIES_ID + YYY + pf_evD.pdf)

Multiple plots of the evolution of box sizes in all non hidden boxes.
(file name structure: CPS + SERIES_ID + YYY + pf_evS.pdf)

Optional outputs

If EACH_RUN_DIGEST = TRUE
Creates and fills DIGEST folder for each run of the composite scenario with all optional outputs of run_isobxr function.
(folder name structure: CPS + SERIES_ID + YYY + XXXX + DIGEST)

If to_CPS_DIGEST_CSVs = TRUE
In the 0_CPS_DIGEST folder, edits csv versions of the whole-composite scenario evD and evS datasets.
(file names structures: CPS + SERIES_ID + YYY + evD.csv and CPS + SERIES_ID + YYY + evS.csv)

See also

Examples

# Example 1: Changing intensity of fluxes
# for more information see tutorial at
# https://ttacail.github.io/isobxr_web/vgn_06_compose_isobxr_tutorial.html#34_Run_outputs

# This is an example using the tutorial files embedded in package data
# It can be run as such.
compose_isobxr(workdir = "/Users/username/Documents/1_ABC_tutorial",
               SERIES_ID = "ABC_change_balance", # name of the series ID
               time_units = c("d", "d"), # in/out time units for pdf plots
               COMPO_MASTER = "0_CPS_MASTER_changing_balance.xlsx",
               plot_HIDE_BOXES_delta = c("SINK"), # hide in delta plots
               plot_HIDE_BOXES_size = c("SOURCE", "SINK"), # hide in size plots
               EACH_RUN_DIGEST = FALSE, # export the DIGEST for each run
               to_CPS_DIGEST_CSVs = TRUE) # export whole model to CSVs

#> ________________________________________________________________________________

#> ℹ workdir: no workdir. You are using the tutorial mode (isobxr embedded tutorial files)

#> ________________________________________________________________________________

#> 🔵 COMPUTING 

#> – The INFINITE boxes are: SOURCE, SINK

#> – Running ana_slvr (balanced finite boxes)

#> [1] -------------------------------------------------- (0% of n = 6)
#> [1] ********------------------------------------------ (17% of n = 6)
#> [1] *****************--------------------------------- (33% of n = 6)
#> [1] *************************------------------------- (50% of n = 6)
#> [1] *********************************----------------- (67% of n = 6)
#> [1] ******************************************-------- (83% of n = 6)
#> [1] ************************************************** (100% of n = 6)

#> ________________________________________________________________________________

#> 🔵 PREPARING RESULTS

#> [1] -------------------------------------------------- (0% of n = 6)
#> [1] ********------------------------------------------ (17% of n = 6)
#> [1] *****************--------------------------------- (33% of n = 6)
#> [1] *************************------------------------- (50% of n = 6)
#> [1] *********************************----------------- (67% of n = 6)
#> [1] ******************************************-------- (83% of n = 6)
#> [1] ************************************************** (100% of n = 6)

#> ________________________________________________________________________________

#> 🔵 WRITING OUTPUTS

#> ________________________________________________________________________________

#> ℹ Run outputs (stored in temporary directory):

#> /var/folders/bj/pyh7pv4922973yvnx5jyvps1f6vt0y/T//RtmpI9TEqP/isobxr_temp_data/
#> ├── 1_LOG.csv
#> └── 3_CPS_ABC_change_balance_001
#>     ├── 0_CPS_DIGEST
#>     │   ├── CPS_ABC_change_balance_001_LOG.csv
#>     │   ├── CPS_ABC_change_balance_001_MASTER.xlsx
#>     │   ├── CPS_ABC_change_balance_001_evD.RDS
#>     │   ├── CPS_ABC_change_balance_001_evD.csv
#>     │   ├── CPS_ABC_change_balance_001_evS.RDS
#>     │   ├── CPS_ABC_change_balance_001_evS.csv
#>     │   ├── CPS_ABC_change_balance_001_p_evDS.pdf
#>     │   ├── CPS_ABC_change_balance_001_pf_evD.pdf
#>     │   └── CPS_ABC_change_balance_001_pf_evS.pdf
#>     ├── CPS_ABC_change_balance_001_0001_IN.Rda
#>     ├── CPS_ABC_change_balance_001_0001_OUT.Rda
#>     ├── CPS_ABC_change_balance_001_0002_IN.Rda
#>     ├── CPS_ABC_change_balance_001_0002_OUT.Rda
#>     ├── CPS_ABC_change_balance_001_0003_IN.Rda
#>     ├── CPS_ABC_change_balance_001_0003_OUT.Rda
#>     ├── CPS_ABC_change_balance_001_0004_IN.Rda
#>     ├── CPS_ABC_change_balance_001_0004_OUT.Rda
#>     ├── CPS_ABC_change_balance_001_0005_IN.Rda
#>     ├── CPS_ABC_change_balance_001_0005_OUT.Rda
#>     ├── CPS_ABC_change_balance_001_0006_IN.Rda
#>     └── CPS_ABC_change_balance_001_0006_OUT.Rda

#> ________________________________________________________________________________

#> ❗ Results were not saved to working directory (set save_run_outputs = TRUE to save results).

#> ℹ You can explore the results with more parameters by using the shinobxr_app() function (requires saved outputs).