Run isobxr stable isotope box model

A function to run the isobxr stable isotope box model, assessing the design of the model and automatically running num_slvr or ana_slvr depending on the conditions.

run_isobxr(
  workdir,
  SERIES_ID,
  flux_list_name,
  coeff_list_name,
  t_lim,
  nb_steps,
  time_units,
  FORCING_RAYLEIGH = NULL,
  FORCING_SIZE = NULL,
  FORCING_DELTA = NULL,
  FORCING_ALPHA = NULL,
  COMPOSITE = FALSE,
  COMPO_SERIES_n = NaN,
  COMPO_SERIES_FAMILY = NaN,
  EXPLORER = FALSE,
  EXPLO_SERIES_n = NaN,
  EXPLO_SERIES_FAMILY = NaN,
  HIDE_PRINTS = FALSE,
  to_DIGEST_DIAGRAMS = TRUE,
  to_DIGEST_evD_PLOT = TRUE,
  to_DIGEST_CSV_XLS = FALSE
)

Arguments

workdir
Working directory of 0_ISOBXR_MASTER.xlsx master file
and where output files will be stored.
(character string)

SERIES_ID
Name of the model series the run belongs to.
It determines the folder in which the output files will be stored.
(character string)

flux_list_name
Name of the list of fluxes and initial box sizes to be used for the run,
calling (by its header name) a single column of the FLUXES sheet of the 0_ISOBXR_MASTER.xlsx file.
(character string)

coeff_list_name
Name of the list of fractionation coefficients to be used for the run,
calling (by its header name) a single column of the COEFFS sheet of the 0_ISOBXR_MASTER.xlsx file.
(character string)

t_lim
Run duration, given in the same time units as the fluxes.
(integer)

nb_steps
Number of calculation steps.
It determines the resolution of the run.
(integer)

time_units
Vector defining the initial time unit (identical to unit used in fluxes),
followed by the time unit used for the graphical output.
Character string, to be selected amongst the following:
micros, ms, s, min, h, d, wk, mo, yr, kyr, Myr, Gyr
e.g., c("d", "yr") to convert days into years

FORCING_RAYLEIGH
OPTIONAL
Dataframe describing the forcing on a fractionation coefficient by a Rayleigh isotope distillation,
as a function of flux intensities and a fundamental fractionation coefficient.
Dataframe formatting details are in isobxr vignette.
Default is NULL.

FORCING_SIZE
OPTIONAL
Dataframe describing the forcing on one or several box sizes (mass of element X).
The newly defined sizes for the given set of boxes overwrite their sizes as previously defined in 0_ISOBXR_MASTER.xlsx file.
Dataframe formatting details are in isobxr vignette.
Default is NULL.

FORCING_DELTA
OPTIONAL
Dataframe describing the forcing on one or several boxes initial isotope composition expressed as delta values.
The newly defined delta values for the given set of boxes overwrite the delta values as previously defined in 0_ISOBXR_MASTER.xlsx file.
Dataframe formatting details are in isobxr vignette.
Default is NULL.

FORCING_ALPHA
OPTIONAL
Dataframe describing the forcing on one or several fractionation coefficients from one reservoir to another.
The newly defined alpha values for the given set of boxes overwrite the alpha values as previously defined in 0_ISOBXR_MASTER.xlsx file.
Dataframe formatting details are in isobxr vignette.
Default is NULL.

COMPOSITE
NOT TO BE USED IN SINGLE RUN
Logical value automatically defined in compose_isobxr.
Default is FALSE.

COMPO_SERIES_n
NOT TO BE USED IN SINGLE RUN
Iteration of the composite run for the given series it belongs to, automatically defined in compose_isobxr.
Default is NaN.

COMPO_SERIES_FAMILY
NOT TO BE USED IN SINGLE RUN
Composite run series family, automatically defined in compose_isobxr.
Default is NaN.

EXPLORER
NOT TO BE USED IN SINGLE RUN
Logical value automatically defined in sweep_steady or sweep_dyn.
Default is FALSE.

EXPLO_SERIES_n
NOT TO BE USED IN SINGLE RUN
Iteration of the sweep run for the given series it belongs to, automatically defined in sweep_steady or sweep_dyn.
Default is NaN.

EXPLO_SERIES_FAMILY
NOT TO BE USED IN SINGLE RUN
Sweep run series family, automatically defined in sweep_steady or sweep_dyn.
Default is NaN.

HIDE_PRINTS
OPTIONAL
Logical value.
Prints outputs details in R console if TRUE.
This parameter does not hide the warnings regarding the automatic update of the run duration in case of the emptying of a box.
Default is FALSE.

to_DIGEST_DIAGRAMS
OPTIONAL
Logical value.
Edits pdf of box model diagram in RUN DIGEST folder if TRUE.
Default is TRUE.

to_DIGEST_evD_PLOT
OPTIONAL
Logical value.
Edits pdf of delta time evolution plot in RUN DIGEST folder if TRUE.
Default is TRUE.

to_DIGEST_CSV_XLS
OPTIONAL
Logical value.
Edits xlsx version of the Rda input file (ending with _IN.xlsx) and all ana_slvr or num_slvr CSV output files in RUN DIGEST folder if TRUE.
Default is FALSE.

Value

If the function is run independently and if directory is not yet existing,
run_isobxr creates and stores all outputs in a SERIES folder located in working directory, with the following name structure:
2_RUN + SERIES_ID

Automatically sets a XXXX run number between 0001 and 9999.
The outputs do not overwrite possible identical previously performed runs.

Stores all run commands in a file with the Rda format.
This file stores all commands used as arguments for solver function (ana_slvr or num_slvr).
(file name structure: SERIES_ID + XXXX + _IN.Rda)
This encompasses the following:

CONSTS_IN: Dataframe documenting the constants (Element, isotopes, reference ratio)

INITIAL_IN: Dataframe documenting the initial conditions of all box sizes (incl. empty boxes)
and all initial delta values.

FLUXES_IN: Dataframe documenting the massic fluxes of element X
between all boxes (structured as a square matrix)

COEFFS_IN: Dataframe documenting the isotopic fractionation coefficients
between all boxes (structured as a square matrix)

BOX_META_IN: Dataframe documenting box metadata
(Box names, initial X masses and delta values, total inward and outward X fluxes in each box,
flux balance for each box, X residence time in each box, max run time before total emptying of each box,
box infinite/finite status, system diagram coordinates)

Stores all outputs in a file with the Rda format.
This file stores all data produced by the function.
(file name structure: SERIES_ID + XXXX + _OUT.Rda)

Updates the general log file, located in general working directory.
(file name: 1_LOG.csv)

Optional outputs, stored in DIGEST folder

A DIGEST folder is created to store all optional outputs of the run_isobxr function.
The DIGEST folder is created in the SERIES folder with the following name structure:
2_RUN + SERIES_ID / SERIES_ID + XXXX + DIGEST

If to_DIGEST_CSV_XLS = TRUE,
creates an INPUT file in the xlsx format stored in DIGEST folder,
containing all run conditions and parameters.
(file name structure: in_0_INPUTS + SERIES_ID + XXXX + .xlsx)

If to_DIGEST_CSV_XLS = TRUE,
stores csv versions of the num_slvr or ana_slvr outputs in DIGEST folder.
See num_slvr or ana_slvr documentation for further details.

If to_DIGEST_DIAGRAMS = TRUE,
edits a Box model diagram of flux (DIAG_FLUX pdf) of element X (mass per time unit) between all boxes.
(file name structure: in_1_DIAG_FLUX + SERIES_ID + XXXX + .pdf)

If to_DIGEST_DIAGRAMS = TRUE,
edits a Box model diagram of isotope fractionation coefficients (DIAG_COEFF pdf) between all boxes.
(file name structure: in_2_DIAG_COEFF + SERIES_ID + XXXX + .pdf)

If to_DIGEST_evD_PLOT = TRUE,
edits a pdf plot of the time dependent evolution of delta values, with a logarithmic x axis time scale.
(file name structure: out_0_PLOT_evD + SERIES_ID + XXXX + .pdf)

See also